+++ Entity/Migrations/AplicationDbContextModelSnapshot.cs
@@ -2738,8 +2738,8 @@ namespace Entity.Migrations
                         .HasColumnType("character varying(100)")
                         .HasColumnName("nation");
 
-                    b.Property<int>("Phone")
-                        .HasColumnType("integer")
+                    b.Property<long>("Phone")
+                        .HasColumnType("bigint")
                         .HasColumnName("phone");
 
                     b.Property<string>("SecondLastName")
@@ -2779,7 +2779,7 @@ namespace Entity.Migrations
                             Identification = 100200300,
                             LastName = "Pérez",
                             Nation = "Colombia",
-                            Phone = 300123456,
+                            Phone = 300123456L,
                             SecondLastName = "García",
                             SecondName = "Andrés",
                             Status = 0
@@ -2794,7 +2794,7 @@ namespace Entity.Migrations
                             Identification = 500600700,
                             LastName = "López",
                             Nation = "Colombia",
-                            Phone = 310987654,
+                            Phone = 310987654L,
                             SecondLastName = "Martínez",
                             SecondName = "Fernanda",
                             Status = 0
@@ -2809,7 +2809,7 @@ namespace Entity.Migrations
                             Identification = 800900100,
                             LastName = "Rodríguez",
                             Nation = "Colombia",
-                            Phone = 320456789,
+                            Phone = 320456789L,
                             SecondLastName = "Hernández",
                             SecondName = "Camilo",
                             Status = 0
@@ -2824,7 +2824,7 @@ namespace Entity.Migrations
                             Identification = 111222333,
                             LastName = "Moreno",
                             Nation = "Colombia",
-                            Phone = 301654987,
+                            Phone = 301654987L,
                             SecondLastName = "Castro",
                             SecondName = "Isabel",
                             Status = 0
@@ -2839,7 +2839,7 @@ namespace Entity.Migrations
                             Identification = 444555666,
                             LastName = "Ramírez",
                             Nation = "Colombia",
-                            Phone = 312789654,
+                            Phone = 312789654L,
                             SecondLastName = "Torres",
                             SecondName = "Esteban",
                             Status = 0
